Chore: split `utils` unit tests into few files
